requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
bcrypt==4.1.3
blinker==1.8.2
blis==0.7.11
cashfree==1.0.1
cashfree-payout==0.0.13
cashfree-pg==4.2.3
cashfree-pg-sdk-python==2.0.1
cashfree_sdk==0.1.3
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
cryptography==42.0.8
cymem==2.0.8
Django==5.0.7
dnspython==2.6.1
docx==0.2.4
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-PyMongo==2.3.0
frozenlist==1.4.1
funcs==0.10.1
geoip2==4.8.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
langcodes==3.4.0
language_data==1.2.0
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
maxminddb==2.6.2
mdurl==0.1.2
mongoengine==0.28.2
multidict==6.0.5
murmurhash==1.0.10
named==1.4.2
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pem==23.1.0
phonenumbers==8.13.40
Pillow==10.0.1
preshed==3.0.9
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
pycparser==2.22
pycryptodome==3.20.0
pydantic==1.10.17
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.8.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-geoip==1.2
pytz==2024.1
PyYAML==6.0.1
razorpay==1.4.2
regex==2024.5.15
requests==2.31.0
requests-file==2.1.0
rfc3986==1.5.0
rich==13.7.1
sentry-sdk==1.32.0
setuptools==70.3.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.0
solus==1.2.2
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
thinc==8.2.5
tk==0.1.0
tldextract==5.1.2
tqdm==4.66.4
typer==0.12.3
typing-aliases==1.10.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.0.7
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.0.3
wraps==0.13.0
wrapt==1.16.0
yarl==1.9.4
